Fix a submission-blocking IndexError: horizon is Float32, not an index
Found by running the whole orchestrator end to end against a structurally faithful
synthetic dataset. All thirteen phases completed; two real bugs fell out.

build_mats casts every feature to Float32 on write and horizon is a feature, so
va["horizon"].to_numpy() is float32 in every matrix. postcal.py and final_assemble.py
both applied the per-horizon calibration as a[np.clip(h, 1, 7)], which raises
"IndexError: arrays used as indices must be of integer (or boolean) type".

postcal crashed, so no calibration was produced at all. final_assemble carries the same
line, so the first time a calibration WAS adopted the submission assembly would have
died -- at the very end of a long run, with the training already spent. Both now cast
explicitly; verified by assembling with the scale-only form and with the affine form.
Nothing else was affected: hsplice, stack, analyze, smooth.horizon_w and run_models'
HFILT all compare or cast rather than index.

Second bug: phase 3b asked layout C for the full nine-experiment table, but C runs only
the finalists, and eval_mix raises on the first missing tag -- so the C table was lost
entirely. Phase 3b now asks C for e1/e4/e8.

Neither bug aborts a run in flight (both steps are `|| true`): such a run finishes and
produces valid submissions, just without the calibration stage and without the C table.
Failed steps never write a marker, so recovery is a pull and a rerun of the same
command -- only those steps and the assembles repeat.

NOTES.md records what the run confirms works, including that validation_c reproduces
the test horizon mix exactly, that select_config reached the same configuration as the
entrant's real run, and that stack.py fitted a four-family ensemble.

+++ final_assemble.py
@@ -75,7 +75,9 @@ if CAL:
     a = np.array([1.0] + [float(x) for x in CAL.split(",")])
     b = np.array([0.0] + [float(x) for x in CALB.split(",")]) if CALB else np.zeros(8)
     assert len(a) == 8 and len(b) == 8, "CALIB/CALIB_B need seven values, one per horizon"
-    j = np.clip(h, 1, 7)
+    # horizon is Float32 in the matrices (build_mats casts every feature), and a
+    # float array cannot index one -- hence the explicit int.
+    j = np.clip(h, 1, 7).astype(int)
     p = k + a[j] * (p - k) + b[j]
     print("calibration " + " ".join(f"h{i}={a[i]:.3f}" + (f"{b[i]:+.3f}" if CALB else "")
                                     for i in range(1, 8)))
